Changes:
added string hashing function since this makes the code a bit cleaner
when using switch case statements.

added core part of the code. Using the switch case statements to 
determine
what part of the code is executed. The statement is determined by
the hash of the string that user gives as input.